+++ tools/api-markdown-documenter/src/index.ts
@@ -66,7 +66,7 @@ export {
 	renderNodeAsMarkdown,
 	renderNodesAsMarkdown,
 } from "./markdown-renderer";
-export { renderApiModelAsMarkdown, renderDocumentsAsMarkdown } from "./MarkdownDocumenter";
+export { renderApiModelAsMarkdown, renderDocumentsAsMarkdown } from "./RenderMarkdown";
 export type { ConfigurationBase } from "./ConfigurationBase";
 export type { Heading } from "./Heading";
 export type { Link, UrlTarget } from "./Link";
